requirement.txt
# Core Dependencies
flask==2.3.3
flask-cors==4.0.0
flask-restful==0.3.10
python-dotenv==1.0.0

# Machine Learning
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
scipy==1.11.2

# NLP
nltk==3.8.1
textblob==0.17.1
spacy==3.6.1

# Deep Learning (optional)
tensorflow==2.13.0
keras==2.13.1

# Data Processing
regex==2023.8.8
emoji==2.8.0

# Database
sqlalchemy==2.0.20
alembic==1.12.0

# API Documentation
flask-swagger-ui==4.11.1
marshmallow==3.20.1

# Monitoring & Logging
loguru==0.7.0

# Security
cryptography==41.0.4
pyjwt==2.8.0

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
httpx==0.24.1

# Deployment
gunicorn==21.2.0
uvicorn==0.23.2

# Utilities
joblib==1.3.2
tqdm==4.66.1
python-dateutil==2.8.2
requests==2.31.0

# Data Visualization (for analysis)
matplotlib==3.7.2
seaborn==0.12.2

# Performance
redis==5.0.0
celery==5.3.1                   